Remove need for lifetime annotations
Because the interaction for serializing and deserializing a mapping is
not generic and part of the trait methods, it makes no sense to have the
Serialize/Deserialize dependency on Crud. By removing this, all the
lifetimes can be removed and I think we can all agree that lifetime
syntax is shit.